feat: refactor footer component into separate DashboardFooter file
