--- src/app/ui/sections/Footer/FooterWrapper.tsx
@@ -1,15 +1,0 @@
-'use client';
-
-import { usePathname } from 'next/navigation';
-
-import { DashboardFooter, Footer } from '@/app/ui/sections/Footer/Footer';
-
-export const FooterWrapper = () => {
-  const pathname = usePathname();
-
-  if (pathname.startsWith('/dashboard')) {
-    return <DashboardFooter />;
-  }
-
-  return <Footer />;
-};
